Fix a race condition which could result in corrupting the PGM-RT state
Nodes were being stored in stack variables and then passed by
reference. Each thread would then create a local copy as soon as
it ran, but if its start was delayed, the reference would no longer
be valid and could point to anything. This is fixed by passing the
node by value.